+++ experimental/packages/exporter-trace-otlp-http/src/platform/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { OTLPTraceExporter } from './node';
+export { createOtlpHttpSpanExporter, OTLPTraceExporter } from './node';