+++ experimental/packages/otlp-grpc-exporter-base/src/index.ts
@@ -3,7 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { convertLegacyOtlpGrpcOptions } from './configuration/convert-legacy-otlp-grpc-options';
+export {
+  convertLegacyOtlpGrpcOptions,
+  convertLegacyOtlpGrpcOptionsWithoutEnv,
+} from './configuration/convert-legacy-otlp-grpc-options';
 export {
   createOtlpGrpcExportDelegate,
   createOtlpGrpcExporterMetrics,